restricted edit,delete buttons and login page

+++ main/views.py
@@ -1,4 +1,6 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
@@ -28,6 +30,7 @@ def room(request, pk):
     return render(request, 'main/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -38,9 +41,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'main/room_form.html', context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("Your not allowed here")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -52,8 +60,12 @@ def updateRoom(request, pk):
     return render(request, 'main/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("Your not allowed here")
 
     if request.method == 'POST':
         room.delete()
